# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Elham Multiplast CRM API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all team members
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a team member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags: [users]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /reports/pipeline:
    get:
      operationId: getPipelineReport
      tags: [reports]
      summary: Deal counts and values by stage
      parameters:
        - name: salesOwnerId
          in: query
          schema:
            type: integer
        - name: unit
          in: query
          schema:
            type: string
        - name: month
          in: query
          schema:
            type: string
        - name: city
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Pipeline breakdown
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PipelineStageCount"
  /reports/by-owner:
    get:
      operationId: getReportByOwner
      tags: [reports]
      summary: Deal stats grouped by sales owner
      parameters:
        - name: month
          in: query
          schema:
            type: string
        - name: unit
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Owner stats
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OwnerStat"
  /reports/by-product:
    get:
      operationId: getReportByProduct
      tags: [reports]
      summary: Deal stats grouped by product
      parameters:
        - name: month
          in: query
          schema:
            type: string
        - name: salesOwnerId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Product stats
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductStat"
  /reports/by-city:
    get:
      operationId: getReportByCity
      tags: [reports]
      summary: Deal stats grouped by city
      parameters:
        - name: month
          in: query
          schema:
            type: string
        - name: salesOwnerId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: City stats
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CityStat"
...
components:
  schemas:
...
    User:
      type: object
      required: [id, name, username, role, colorCode, unit]
      properties:
        id:
          type: integer
        name:
          type: string
        username:
          type: string
        role:
          type: string
          enum: [admin, sales]
        colorCode:
          type: string
        unit:
          type: string
          enum: [Himatnagar, Surat, Rajkot, All]
        createdAt:
          type: string
    UserInput:
      type: object
      required: [name, username, password, role, colorCode, unit]
      properties:
        name:
          type: string
        username:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [admin, sales]
        colorCode:
          type: string
        unit:
          type: string
          enum: [Himatnagar, Surat, Rajkot, All]
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        username:
          type: string
        password:
          type: ["string", "null"]
        role:
          type: string
          enum: [admin, sales]
        colorCode:
          type: string
        unit:
          type: string
...
    PipelineStageCount:
      type: object
      required: [stage, count, totalValue]
      properties:
        stage:
          type: string
        count:
          type: integer
        totalValue:
          type: number
        probability:
          type: integer
    OwnerStat:
      type: object
      required: [userId, userName, colorCode, totalDeals, wonDeals, lostDeals, totalWonValue]
      properties:
        userId:
          type: integer
        userName:
          type: string
        colorCode:
          type: string
        totalDeals:
          type: integer
        wonDeals:
          type: integer
        lostDeals:
          type: integer
        activeDeals:
          type: integer
        totalWonValue:
          type: number
    ProductStat:
      type: object
      required: [productId, productName, totalQuantity, totalValue, dealCount]
      properties:
        productId:
          type: integer
        productName:
          type: string
        productCode:
          type: string
        totalQuantity:
          type: number
        totalValue:
          type: number
        dealCount:
          type: integer
    CityStat:
      type: object
      required: [city, totalDeals, wonDeals, totalWonValue]
      properties:
        city:
          type: string
        totalDeals:
          type: integer
        wonDeals:
          type: integer
        totalWonValue:
          type: number
...