# lib/api-zod/src/generated/types/ownerStat.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Elham Multiplast CRM API
 * OpenAPI spec version: 0.1.0
 */

export interface OwnerStat {
  userId: number;
  userName: string;
  colorCode: string;
  totalDeals: number;
  wonDeals: number;
  lostDeals: number;
  activeDeals?: number;
  totalWonValue: number;
}