# lib/api-zod/src/generated/types/userUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Elham Multiplast CRM API
 * OpenAPI spec version: 0.1.0
 */
import type { UserUpdateRole } from './userUpdateRole';

export interface UserUpdate {
  name?: string;
  username?: string;
  /** @nullable */
  password?: string | null;
  role?: UserUpdateRole;
  colorCode?: string;
  unit?: string;
}